+++ app/src/main/java/com/github/bumblebee202111/doubean/data/repository/SubjectCommonRepository.kt
@@ -1,12 +1,15 @@
 package com.github.bumblebee202111.doubean.data.repository
 
 import com.github.bumblebee202111.doubean.model.AppResult
+import com.github.bumblebee202111.doubean.model.subjects.RecommendSubject
 import com.github.bumblebee202111.doubean.model.subjects.SubjectModule
 import com.github.bumblebee202111.doubean.model.subjects.SubjectReviewList
 import com.github.bumblebee202111.doubean.model.subjects.SubjectType
 import com.github.bumblebee202111.doubean.network.ApiService
 import com.github.bumblebee202111.doubean.network.model.NetworkSubjectModules
 import com.github.bumblebee202111.doubean.network.model.NetworkSubjectReviewList
+import com.github.bumblebee202111.doubean.network.model.subject.NetworkRecommend
+import com.github.bumblebee202111.doubean.network.model.subject.toRecommendSubject
 import com.github.bumblebee202111.doubean.network.model.toNetworkSubjectType
 import com.github.bumblebee202111.doubean.network.model.toSubjectModules
 import com.github.bumblebee202111.doubean.network.model.toSubjectReviewList
@@ -16,6 +19,29 @@ import javax.inject.Singleton
 
 @Singleton
 class SubjectCommonRepository @Inject constructor(private val service: ApiService) {
+
+    suspend fun getSubjectRelatedItems(
+        subjectType: SubjectType,
+        subjectId: String,
+    ): AppResult<List<RecommendSubject>> {
+        return makeApiCall(
+            apiCall = {
+                service.getSubjectRelatedItems(
+                    subjectType = subjectType.toNetworkSubjectType().value,
+                    subjectId = subjectId
+                )
+            },
+            mapSuccess = { networkRecommendations ->
+                networkRecommendations.subjects.mapNotNull { recommend ->
+                    when (recommend) {
+                        is NetworkRecommend.Subject -> recommend.subject.toRecommendSubject()
+                        else -> null
+                    }
+                }
+            }
+        )
+    }
+
     suspend fun getSubjectReviews(
         subjectType: SubjectType,
         subjectId: String,