+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/subjects/movie/MovieUiState.kt
@@ -2,6 +2,7 @@ package com.github.bumblebee202111.doubean.feature.subjects.movie
 
 import com.github.bumblebee202111.doubean.model.PhotoList
 import com.github.bumblebee202111.doubean.model.subjects.MovieDetail
+import com.github.bumblebee202111.doubean.model.subjects.RecommendSubject
 import com.github.bumblebee202111.doubean.model.subjects.SubjectInterestWithUserList
 import com.github.bumblebee202111.doubean.model.subjects.SubjectReviewList
 
@@ -10,9 +11,11 @@ sealed interface MovieUiState {
         val movie: MovieDetail,
         val interests: SubjectInterestWithUserList,
         val photos: PhotoList,
+        val recommendations: List<RecommendSubject>,
         val reviews: SubjectReviewList,
         val isLoggedIn: Boolean,
-    ) : MovieUiState
+
+        ) : MovieUiState
     data object Loading : MovieUiState
     data object Error : MovieUiState
 }